+++ packages/openai/src/embedding/openai-embedding-options.ts
@@ -11,15 +11,15 @@ export const openaiEmbeddingProviderOptions = lazySchema(() =>
   zodSchema(
     z.object({
       /**
-The number of dimensions the resulting output embeddings should have.
-Only supported in text-embedding-3 and later models.
-   */
+       * The number of dimensions the resulting output embeddings should have.
+       * Only supported in text-embedding-3 and later models.
+       */
       dimensions: z.number().optional(),
 
       /**
-A unique identifier representing your end-user, which can help OpenAI to
-monitor and detect abuse. Learn more.
-*/
+       * A unique identifier representing your end-user, which can help OpenAI to
+       * monitor and detect abuse. Learn more.
+       */
       user: z.string().optional(),
     }),
   ),