+++ packages/provider/src/language-model/v2/language-model-v2-text.ts
@@ -1,13 +1,13 @@
 import { SharedV2ProviderMetadata } from '../../shared/v2/shared-v2-provider-metadata';
 
 /**
-Text that the model has generated.
+ * Text that the model has generated.
  */
 export type LanguageModelV2Text = {
   type: 'text';
 
   /**
-The text content.
+   * The text content.
    */
   text: string;