+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ai-sdk/azure
 
+## 3.0.21
+
+### Patch Changes
+
+- Updated dependencies [2c70b90]
+  - @ai-sdk/openai@3.0.21
+
+## 3.0.20
+
+### Patch Changes
+
+- Updated dependencies [462ad00]
+  - @ai-sdk/provider-utils@4.0.10
+  - @ai-sdk/openai@3.0.20
+
+## 3.0.19
+
+### Patch Changes
+
+- 04c89b1: Provide Responses API providerMetadata types at the message / reasoning level.
+
+  - Export the following types for use in client code:
+    - `OpenaiResponsesProviderMetadata`
+    - `OpenaiResponsesReasoningProviderMetadata`
+    - `AzureResponsesProviderMetadata`
+    - `AzureResponsesReasoningProviderMetadata`
+
+- Updated dependencies [04c89b1]
+  - @ai-sdk/openai@3.0.19
+
 ## 3.0.18
 
 ### Patch Changes