+++ packages/fireworks/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/fireworks
 
+## 2.0.22
+
+### Patch Changes
+
+- Updated dependencies [462ad00]
+  - @ai-sdk/provider-utils@4.0.10
+  - @ai-sdk/openai-compatible@2.0.21
+
+## 2.0.21
+
+### Patch Changes
+
+- Updated dependencies [a1a0175]
+  - @ai-sdk/openai-compatible@2.0.20
+
+## 2.0.20
+
+### Patch Changes
+
+- Updated dependencies [6900916]
+  - @ai-sdk/openai-compatible@2.0.19
+
 ## 2.0.19
 
 ### Patch Changes