+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ai-sdk/gateway
 
+## 3.0.26
+
+### Patch Changes
+
+- a8be296: chore(provider/gateway): update gateway model settings files
+
+## 3.0.25
+
+### Patch Changes
+
+- 15a78c7: chore(provider/gateway): update gateway model settings files
+
+## 3.0.24
+
+### Patch Changes
+
+- Updated dependencies [462ad00]
+  - @ai-sdk/provider-utils@4.0.10
+
+## 3.0.23
+
+### Patch Changes
+
+- cbf1704: chore(provider/gateway): update gateway model settings files
+
 ## 3.0.22
 
 ### Patch Changes