+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @ai-sdk/google-vertex
 
+## 4.0.34
+
+### Patch Changes
+
+- Updated dependencies [97b3ebb]
+  - @ai-sdk/google@3.0.16
+
+## 4.0.33
+
+### Patch Changes
+
+- Updated dependencies [2445da4]
+  - @ai-sdk/anthropic@3.0.28
+
+## 4.0.32
+
+### Patch Changes
+
+- Updated dependencies [572ea12]
+  - @ai-sdk/anthropic@3.0.27
+
+## 4.0.31
+
+### Patch Changes
+
+- Updated dependencies [2c70b90]
+  - @ai-sdk/anthropic@3.0.26
+  - @ai-sdk/google@3.0.15
+
+## 4.0.30
+
+### Patch Changes
+
+- Updated dependencies [0bb9bcd]
+  - @ai-sdk/anthropic@3.0.25
+
+## 4.0.29
+
+### Patch Changes
+
+- Updated dependencies [462ad00]
+  - @ai-sdk/provider-utils@4.0.10
+  - @ai-sdk/anthropic@3.0.24
+  - @ai-sdk/google@3.0.14
+
 ## 4.0.28
 
 ### Patch Changes