+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ai-sdk/openai
 
+## 3.0.21
+
+### Patch Changes
+
+- 2c70b90: chore: update provider docs
+
+## 3.0.20
+
+### Patch Changes
+
+- Updated dependencies [462ad00]
+  - @ai-sdk/provider-utils@4.0.10
+
+## 3.0.19
+
+### Patch Changes
+
+- 04c89b1: Provide Responses API providerMetadata types at the message / reasoning level.
+
+  - Export the following types for use in client code:
+    - `OpenaiResponsesProviderMetadata`
+    - `OpenaiResponsesReasoningProviderMetadata`
+    - `AzureResponsesProviderMetadata`
+    - `AzureResponsesReasoningProviderMetadata`
+
 ## 3.0.18
 
 ### Patch Changes